requirements.txt
google-adk
google-cloud-firestore
requests
python-dotenv
